requirements.txt
typer>=0.9
pyfaidx>=0.8
numpy
scipy
fisher
pyx
pypdfium2
Pillow

